Refs 9393. Take IPayloadPool interface outside
